+++ extensions/penpot/src/search-projects.tsx
@@ -1,7 +1,8 @@
 import { ActionPanel, Action, Icon, List } from "@raycast/api";
 import type { Project } from "../types";
-import { usePenpotFetch } from "../utils";
+import { formatDate, usePenpotFetch } from "../utils";
 import { useEffect, useState } from "react";
+import { groupByTeam } from "./utils";
 
 export default function Command() {
   const { isLoading, data: projects, revalidate } = usePenpotFetch<Project[]>("/get-all-projects", {});
@@ -17,6 +18,9 @@ export default function Command() {
       ),
     );
   }, [searchText]);
+
+  const projectsPerTeam = groupByTeam(projects || []);
+  const teamNames = Object.keys(projectsPerTeam);
 
   return (
     <List
@@ -35,6 +39,33 @@ export default function Command() {
         </ActionPanel>
       }
     >
+      {teamNames.map((teamName) => (
+        <List.Section key={teamName} title={teamName}>
+          {projectsPerTeam[teamName]
+            .filter(
+              (project) =>
+                project.name.toLowerCase().includes(searchText.toLowerCase()) ||
+                project.teamName.toLowerCase().includes(searchText.toLowerCase()),
+            )
+            .map((project) => (
+              <List.Item
+                key={project.id}
+                title={project.name}
+                icon={Icon.BlankDocument}
+                accessories={[
+                  { date: new Date(project.modifiedAt), tooltip: `Modified ${formatDate(project.modifiedAt)}` },
+                ]}
+                actions={
+                  <ActionPanel>
+                    <Action.OpenInBrowser
+                      url={`https://design.penpot.app/#/dashboard/team/${project.teamId}/projects/${project.id}`}
+                    />
+                  </ActionPanel>
+                }
+              />
+            ))}
+        </List.Section>
+      ))}
       {filteredProjects.map((project) => (
         <List.Item
           key={project.id}